+++ src/components/avatar/AvatarInitials.js
@@ -1,6 +1,6 @@
 import React, { PureComponent } from 'react';
 import PropTypes from 'prop-types';
-import Box, { pickBoxProps } from '../box';
+import Box from '../box';
 import cx from 'classnames';
 import theme from './theme.css';
 import { Heading4 } from '../typography';
@@ -27,10 +27,8 @@ class AvatarInitials extends PureComponent {
       className,
     );
 
-    const boxProps = pickBoxProps(others);
-
     return (
-      <Box className={avatarClassNames} {...boxProps} data-teamleader-ui="avatar-initials">
+      <Box className={avatarClassNames} {...others} data-teamleader-ui="avatar-initials">
         <Heading4 className={theme['content']}>{this.getInitials()}</Heading4>
       </Box>
     );
@@ -45,12 +43,13 @@ AvatarInitials.propTypes = {
   /** The shape of the avatar. */
   shape: PropTypes.oneOf(['circle', 'rounded']),
   /** The size of the avatar. */
-  size: PropTypes.oneOf(['tiny', 'small', 'medium']),
+  size: PropTypes.oneOf(['tiny', 'small', 'medium', 'large']),
   /** The name for in the avatar. */
   name: PropTypes.string,
 };
 
 AvatarInitials.defaultProps = {
+  color: 'neutral',
   shape: 'circle',
   size: 'medium',
   name: 'Michael Scott',