+++ src/components/label/Label.js
@@ -2,22 +2,10 @@ import React, { PureComponent } from 'react';
 import PropTypes from 'prop-types';
 import Box from '../box';
 import { TextBody, TextDisplay, TextSmall } from '../typography';
-import theme from './theme.css';
-import cx from 'classnames';
 
 export default class Label extends PureComponent {
   render() {
-    const {
-      children,
-      connectedLeft,
-      connectedRight,
-      className,
-      inverse,
-      helpText,
-      required,
-      size,
-      ...others
-    } = this.props;
+    const { children, connectedLeft, connectedRight, inverse, helpText, required, size, ...others } = this.props;
 
     const childProps = {
       inverse,
@@ -25,18 +13,10 @@ export default class Label extends PureComponent {
       size,
     };
 
-    const classNames = cx(
-      theme['label'],
-      {
-        [theme['is-inverse']]: inverse,
-      },
-      className,
-    );
-
     const Element = size === 'large' ? TextDisplay : TextBody;
 
     return (
-      <Box element="label" marginBottom={3} className={classNames} {...others}>
+      <Box display="block" element="label" marginBottom={3} {...others}>
         {React.Children.map(children, child => {
           if (typeof child !== 'string') {
             return React.cloneElement(child, { ...childProps, ...child.props });
@@ -76,7 +56,6 @@ export default class Label extends PureComponent {
 }
 
 Label.propTypes = {
-  className: PropTypes.string,
   children: PropTypes.oneOfType([PropTypes.element, PropTypes.string, PropTypes.array]),
   connectedLeft: PropTypes.element,
   connectedRight: PropTypes.element,