+++ src/components/progressTracker/ProgressStep.js
@@ -8,33 +8,33 @@ import { TextSmall } from '../typography';
 
 class ProgressStep extends PureComponent {
   render() {
-    const { label, active, completed, color } = this.props;
-    const classNames = cx(theme['step'], theme[color], {
+    const { label, active, completed, onClick } = this.props;
+    const classNames = cx(theme['step'], {
       [theme['is-active']]: active,
       [theme['is-completed']]: completed,
+      [theme['is-clickable']]: onClick,
     });
     return (
       <Box className={classNames}>
         <TextSmall className={theme['step-label']}>{label}</TextSmall>
-        <span className={theme['status-bullet']} />
+        <span className={theme['status-bullet']} onClick={onClick} />
       </Box>
     );
   }
 }
 
 ProgressStep.propTypes = {
-  /** Color theme of the progress tracker. */
-  color: PropTypes.oneOf(['neutral', 'mint', 'aqua', 'violet', 'gold', 'ruby']),
   /** The label for the progress step */
   label: PropTypes.string.isRequired,
   /** Whether or not the step is active */
   active: PropTypes.bool.isRequired,
   /** Whether or not the step has been completed */
   completed: PropTypes.bool.isRequired,
+  /** Callback function that is fired when the progress step is clicked */
+  onClick: PropTypes.func,
 };
 
 ProgressStep.defaultProps = {
-  color: 'neutral',
   active: false,
   completed: false,
 };