+++ stories/avatars.js
@@ -5,7 +5,7 @@ import { Avatar, AvatarInitials, AvatarStack, Bullet, Counter, TextBody, Tooltip
 import avatars from './static/data/avatar';
 
 const directions = ['horizontal', 'vertical'];
-const sizes = ['tiny', 'small', 'medium'];
+const sizes = ['tiny', 'small', 'medium', 'large'];
 const shapes = [null, 'circle', 'rounded'];
 const colors = ['teal', 'neutral', 'mint', 'violet', 'ruby', 'gold', 'aqua'];
 
@@ -39,12 +39,14 @@ storiesOf('Avatars', module)
       inverse={boolean('Inverse', false)}
       size={select('Size', sizes, 'medium')}
     >
-      {avatars.map(({ image }, index) => <Avatar key={index} image={image} />)}
+      {avatars.map(({ image }, index) => (
+        <Avatar key={index} image={image} />
+      ))}
     </AvatarStack>
   ))
   .add('With bullet', () => (
     <Avatar image={avatars[0].image} size={select('Size', sizes, 'medium')}>
-      <Bullet color="ruby" />
+      <Bullet borderColor="neutral" borderTint="lightest" color="ruby" />
     </Avatar>
   ))
   .add('With counter', () => (