+++ stories/select.js
@@ -6,6 +6,7 @@ import { customOptions, groupedOptions, options } from './static/data/select';
 
 const sizes = ['small', 'medium', 'large'];
 
+/* eslint react/prop-types: 0 */
 const CustomOption = ({ children, data, innerProps, isFocused, isSelected, isDisabled }) => {
   const boxStyles = {
     backgroundColor: isFocused ? '#e4e4e6' : isSelected ? '#82828c' : '#fff',